Replace SyncBufferRepository with CentralSyncBufferRepository
Renames `SyncBufferRepository` to `CentralSyncBufferRepository` and
adds/removes missing/out-of-scope methods (see
https://github.com/openmsupply/remote-server/issues/278).
