+++ src/database/schema/mod.rs
@@ -24,7 +24,7 @@ pub enum DatabaseRow {
     TransactLine(TransactLineRow),
     UserAccount(UserAccountRow),
     SyncOut(SyncOutRow),
-    SyncBuffer(SyncBufferRow),
+    CentralSyncBuffer(CentralSyncBufferRow),
 }
 
 pub use central_sync_buffer::CentralSyncBufferRow;
